add typescript, restructurize files and folders, add npm script to compile ts to js and start nodemon server
