+++ .gitIgnore
@@ -1,2 +1,3 @@
 .env
 /node_modules
+/build
