+++ src/app.js
@@ -2,10 +2,10 @@ const express = require("express")
 const app = express()
 const cors = require("cors")
 const mongoose = require("mongoose")
-const mainRouter = require("./modules/routes/router")
+const mainRouter = require("./routes/router")
 
 const bodyparser = require('body-parser')
-mongoose.connect('mongodb://10.81.7.29:27017/')
+mongoose.connect('mongodb+srv://murauskasnorbertas:slaptas@cluster0.0pdeqi5.mongodb.net/ChecklistDB')
     .then(() => {
         console.log('CONNECTED OK')
     }).catch(e => {
