+++ src/middleware/middle.js
@@ -1,6 +1,6 @@
 const isEmail =             require("is-email")
 const sendRes =             require("../modules/UniversalRes")
-const UserRegisterSchema =  require("../../shemas/UserRegisterSchema")
+const UserRegisterSchema =  require("../shemas/UserRegisterSchema")
 const jwt =                 require("jsonwebtoken");
 const MongoClient =         require('mongodb').MongoClient;
 const client =              new MongoClient('mongodb://10.81.7.29:27017/');
