+++ src/signatureLogo.png
